+++ integrations/slack/src/handlers/index.ts
@@ -1,2 +1,5 @@
 export * from './search';
 export * from './lens';
+export * from './actions';
+export * from './commands';
+export * from './events';
